+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/State.kt
@@ -1,0 +1,5 @@
+package dev.inmo.micro_utils.fsm.common
+
+interface State {
+    val context: Any
+}