+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/StatesHandler.kt
@@ -1,0 +1,5 @@
+package dev.inmo.micro_utils.fsm.common
+
+fun interface StatesHandler<I : State> {
+    suspend fun StatesMachine.handleState(state: I): State?
+}