+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     * Add functionality for multiplatform working with files:
         * Main class for files `MPPFile`
         * Inline class for filenames work encapsulation `FileName`
+* `FSM`
+    * Module inited and in preview state
 
 ## 0.5.12